more removal of swd for pass and make a stillSpeakerTriplet calculation method
Due to HDC being fast, we should still be fine using SWD math for robot angle. But with regards to pivot angles + shooter speeds, I don't think it is worth it as of now to be using.
The addition of this method and removal of swd calc in teleop makes it much easier to shoot the moment the bot stops moving, which is all of teleop.